requirements.txt
# OverconfidenceLens — Python 3.11
# Install: pip install -r requirements.txt

# LLM API
openai>=1.30.0

# Config & environment
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Code execution sandbox (SETUP.md verify: import docker)
docker>=7.0.0

# Numerics & statistics (src/statistical_tests.py)
numpy>=1.26.0
scipy>=1.11.0

# Plotting (Phase 4 — ResultVisualizer)
matplotlib>=3.8.0

# Referenced in module stubs (linguistic_annotator: Cohen's kappa; statistical_tests: McNemar)
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Tests (SETUP.md: pytest tests/ -v)
pytest>=7.4.0
